requirements.txt
python-telegram-bot[asyncio]~=20.6
sqlalchemy~=2.0.25
asyncpg~=0.27.0
aiohttp~=3.8.5
aiofiles~=23.2.1
uvloop ; sys_platform != "win32"
python-dotenv~=1.0.1
redis~=5.0.1
